requirements.txt
openai==1.72.0
pytest-asyncio==0.26.0
pandas==2.2.3
tabulate==0.9.0
dotenv==0.9.9
pydantic-settings==2.8.1
PyYAML